Make Promise.Then and Promise.Catch propagate result properly
Currently, `Then` and `Catch` are declared to return a single value,
which makes raising errors in `Then` or recovering in `Catch`
impossible.  Fix this by adjusting the signature of `Then` and `Catch`
handlers to accept functions returning a pair of values.  An error
returned from `Then` is then passed to the next `Catch`, and if `Catch`
returns a non-error value, it is taken as the resolution for the promise
chain.  This conforms to the behavior of JavaScript Promise objects.